(feat): add hh-modules
